feat: module definitions
